requirements.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
djangorestframework==3.11.0
fuzzywuzzy==0.18.0
google-api-core==1.20.1
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
lml==0.0.9
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyexcel==0.6.4
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
texttable==1.6.2
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
